Accept :success and :failure opts in mock_rpc
Previously to invoke the on_success cb on a client
call you had to provide a :response. The default
key is now to provide a :success object. We will
still pull the :response object as well for
backwards compatability.

Similar to :success/:response, we now accept
:failure/:error as synonymous options, with
:failure being the preferred option as it is more
closely tied to the on_failure nomenclature.