Upgrade Protobuf::RSpec to work with Protobuf 3.0
Up the minimum Protobuf version to 3.0.0.alpha. With the dispatcher now
part of the middleware stack and the service now being initialized with
an env object, rework the local_rpc to invoke the service directly.

Also add a few more helpers for testing the full RPC stack: rpc,
rpc_env, wrapped_request and update the README accordingly.